requirements.txt
PyPDF2 == 3.0.1
SpeechRecognition  == 3.11.0
Django == 5.1.3
moviepy == 1.0.3
python-docx == 1.1.2
pytube == 15.0.0
django